Moving backfiller to separate module
